Add heuristic tests

+++ src/main/java/search_algorithms/HeuristicSearch.java
@@ -22,12 +22,12 @@ public class HeuristicSearch<T extends Comparable<T>> extends Search<T> {
         childParent = new HashMap<>();
     }
 
-    double manhattanDistance(T state) {
-        int stateInt = (Integer) state;
+    int manhattanDistance(T state) {
+        long stateInt = (Long) state;
 
         int manhattanDistance = 0;
         for (int i = 8; i >= 0; i--) {
-            int stateDigit = stateInt % 10;
+            int stateDigit = (int) stateInt % 10;
             int stateRow = i / 3;
             int stateCol = i % 3;
             int goalRow = stateDigit / 3;
@@ -39,11 +39,11 @@ public class HeuristicSearch<T extends Comparable<T>> extends Search<T> {
     }
 
     double EuclideanDistance(T state) {
-        int stateInt = (Integer) state;
+        long stateInt = (Long) state;
 
         double euclideanDistance = 0;
         for (int i = 8; i >= 0; i--) {
-            int stateDigit = stateInt % 10;
+            int stateDigit = (int) (stateInt % 10);
             int stateRow = i / 3;
             int stateCol = i % 3;
             int goalRow = stateDigit / 3;
@@ -59,7 +59,7 @@ public class HeuristicSearch<T extends Comparable<T>> extends Search<T> {
     // now we will implement A* search algorithm
     @Override
     public boolean search(T initialState, T goalState) {
-        PriorityQueue<Pair<Double, T>> frontier = new PriorityQueue<>((a, b) -> Integer.compare(a.getKey().compareTo(b.getKey()), 0));
+        PriorityQueue<Pair<Double, T>> frontier = new PriorityQueue<>((a, b) -> Long.compare(a.getKey().compareTo(b.getKey()), 0));
         double Fn = this.isManhattan ? manhattanDistance(initialState) : EuclideanDistance(initialState);
         frontier.add(Pair.of(Fn, initialState));
         childParent.put(initialState, Pair.of(0.0, initialState));
@@ -90,19 +90,13 @@ public class HeuristicSearch<T extends Comparable<T>> extends Search<T> {
     }
 
     @Override
-    public T getParent(T state) {
-        return childParent.get(state).getRight();
-    }
+    public T getParent(T state) { return childParent.get(state).getRight(); }
 
     @Override
-    public Integer getDepth() {
-        return this.toGoalPathCost;
-    }
+    public Integer getDepth() { return this.toGoalPathCost; }
 
     @Override
-    public Integer getCostOfPath() {
-        return childParent.get(initialState).getLeft().intValue();
-    }
+    public Integer getCostOfPath() { return childParent.get(initialState).getLeft().intValue(); }
 
     @Override
     public Integer getNodesExpanded() {
